- fixed some variable mistakes

+++ SyZaFi/DBConnection.cs
@@ -204,7 +204,7 @@ namespace SyZaFi
             return data;
         }
 
-        public void InsertNewEmployee(string firstName, string lastName, string fathersName, string mothersName, string birthday, string placeOfResidence, string correspondenceAddress, string education, string historyOfEmployment, int pesel, string childrensNames, string childrensBirthdays)
+        public void InsertNewEmployee(string firstName, string lastName, string fathersName, string mothersName, string birthday, string placeOfResidence, string correspondenceAddress, string education, string historyOfEmployment, long pesel, string childrensNames, string childrensBirthdays)
         {
             string query = "INSERT INTO employees (firstName, lastName, fathersName, mothersName, birthday, placeOfResidence, correspondenceAddress, education, historyOfEmployment, pesel, childrensNames, childrensBirthdays)" +
                 " VALUES('" + firstName + "', '" + lastName + "', '" + fathersName + "', '" + mothersName + "', '" + birthday + "', '" + placeOfResidence + "', '" + correspondenceAddress + "', '" + education + "', '" + historyOfEmployment + "', '" + pesel + "', '" + childrensNames + "', '" + childrensBirthdays + "')";
@@ -681,7 +681,7 @@ namespace SyZaFi
             }
         }
 
-        public void InsertNewCompany(string companyName, int companyNIP, long companyREGON, int companyKRS)
+        public void InsertNewCompany(string companyName, long companyNIP, long companyREGON, long companyKRS)
         {
             string query = "INSERT INTO companydetails (companyName, companyNIP, companyREGON, companyKRS)" +
                 " VALUES('" + companyName + "', '" + companyNIP + "', '" + companyREGON + "', '" + companyKRS + "')";
